style photo select done

+++ app/src/main/java/com/davis/a9concept/SelectSizeActivity.java
@@ -5,9 +5,7 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
-import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -32,7 +30,7 @@ public class SelectSizeActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 Intent intent = new Intent(SelectSizeActivity.this, DynamicTabsActivity.class);
-                intent.putExtra("size","1");
+                intent.putExtra("size", "1");
                 startActivity(intent);
             }
         });
@@ -40,7 +38,7 @@ public class SelectSizeActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 Intent intent = new Intent(SelectSizeActivity.this, DynamicTabsActivity.class);
-                intent.putExtra("size","2");
+                intent.putExtra("size", "2");
                 startActivity(intent);
             }
         });
@@ -48,7 +46,7 @@ public class SelectSizeActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 Intent intent = new Intent(SelectSizeActivity.this, DynamicTabsActivity.class);
-                intent.putExtra("size","3");
+                intent.putExtra("size", "3");
                 startActivity(intent);
             }
         });
@@ -65,23 +63,23 @@ public class SelectSizeActivity extends AppCompatActivity {
         RelativeLayout frame_three = findViewById(R.id.frame_rl_three);
 
 
-        frame_one.getLayoutParams().width =  (int) (rl_header.getWidth()/1.3);
-        frame_two.getLayoutParams().width =  (int) (rl_header.getWidth()/1.3);
-        frame_three.getLayoutParams().width =  (int) (rl_header.getWidth()/1.3);
+        frame_one.getLayoutParams().width = (int) (rl_header.getWidth() / 1.3);
+        frame_two.getLayoutParams().width = (int) (rl_header.getWidth() / 1.3);
+        frame_three.getLayoutParams().width = (int) (rl_header.getWidth() / 1.3);
 
-        int height_1 =   frame_one.getLayoutParams().width;
-        int height_2 =   frame_two.getLayoutParams().width;
-        int height_3 =   frame_three.getLayoutParams().width;
-        Log.d("width","set width" + frame_two.getLayoutParams().width);
+        int height_1 = frame_one.getLayoutParams().width;
+        int height_2 = frame_two.getLayoutParams().width;
+        int height_3 = frame_three.getLayoutParams().width;
+        Log.d("width", "set width" + frame_two.getLayoutParams().width);
 
         frame_one.requestLayout();
         frame_two.requestLayout();
         frame_three.requestLayout();
 
-        frame_one.getLayoutParams().height =  (int) (height_1/1.5);
-        frame_two.getLayoutParams().height =  (int) (height_2/2);
-        Log.d("width","get width" + frame_two.getWidth());
-        frame_three.getLayoutParams().height =  height_3;
+        frame_one.getLayoutParams().height = (int) (height_1 / 1.5);
+        frame_two.getLayoutParams().height = height_2 / 2;
+        Log.d("width", "get width" + frame_two.getWidth());
+        frame_three.getLayoutParams().height = height_3;
 
         frame_one.requestLayout();
         frame_two.requestLayout();
